Rename python to local
